style(notification): change package and file name
~ form api to handler package
~ from handler.go to notifications.go
This is done so I can add the preferences.go handler here, with less ambiguity
