+++ services/notification/transport/handler/healthcheck.go
@@ -1,4 +1,4 @@
-package api
+package handler
 
 import (
 	"encoding/json"
